WIP Try to get createStore tests fixed
It seems the fake indexedDB doesn't reset between tests so each call to createStore initializes with the previous stores feed data. Trying to call `deleteDB` fails once we've opened a feed. I attempted to return to the feed from `createStore` so we could try to close it in the tests. This just results in errors from the fake indexedDB.