Restore feed writing to middleware
This requires using the proxyReducer to predict which documents will be changed so we can cache them before the changes are applied. This was the best way I could think of to keep us from cloning the docSet and breaking Automerge change logging.

Note: For some reason the offline tests cause timeouts for the one or more of the "connected" tests. Will figure that out later