toolbar/tsconfig: remove unwanted `include` entry
This is what was making it impossible for Diego to build a cleanly cloned repo - it structured the output from `cevitxe-toolbar` into `src`, `@types`, and `workers` folders, so the file `dist/index.js` that `package.json` pointed to didn't exist. Herb wasn't seeing this because a previous `dist/index.js` was still there, so that's a separate issue: we need to remove any existing build output before building so that each build is clean.